fix inconsistencies in client modules to align them to angular conventions better
