+++ src/client/app/+detail/detail.template.html
@@ -1,0 +1,2 @@
+<h1>Hello from Detail</h1>
+<router-outlet></router-outlet>
