+++ src/client/app/about/about.template.html
@@ -1,0 +1,7 @@
+<h1>About</h1>
+<div>
+  <h3>
+    david.r.niciforovic@gmail.com
+  </h3>
+</div>
+<pre>this.localState = {{ localState | json }}</pre>
