+++ src/client/app/admin/admin.template.html
@@ -1,0 +1,10 @@
+<h3>Admin Dash</h3>
+<nav>
+  <a routerLink="./" routerLinkActive="active"
+     [routerLinkActiveOptions]="{ exact: true }">
+    Dashboard
+  </a>
+  <a routerLink="./issues" routerLinkActive="active">Manage Issues</a>
+  <a routerLink="./users" routerLinkActive="active">Manage Users</a>
+</nav>
+<router-outlet></router-outlet>
