+++ src/client/app/admin/dash/dash.component.ts
@@ -1,9 +1,7 @@
 import { Component } from '@angular/core';
 
 @Component({
-  template: `
-    <p>Dashboard</p>
-  `
+  templateUrl: './dash.template.html'
 })
 
 export class AdminDashComponent { }
