+++ src/client/app/admin/dash/dash.template.html
@@ -1,0 +1,1 @@
+<p>Dashboard</p>
