+++ src/client/app/admin/manage-issues/manage-issues.template.html
@@ -1,0 +1,1 @@
+<p>Manage your issues here</p>
