+++ src/client/app/admin/manage-users/manage-users.template.html
@@ -1,0 +1,1 @@
+<p>Manage your users here</p>
