+++ src/client/app/app.component.html
@@ -1,0 +1,66 @@
+<nav>
+  <span>
+    <a [routerLink]=" ['./'] ">
+      Index
+    </a>
+  </span>
+  |
+  <span>
+    <a [routerLink]=" ['./home'] ">
+      Home
+    </a>
+  </span>
+  |
+  <span>
+    <a [routerLink]=" ['./detail'] ">
+      Detail
+    </a>
+  </span>
+  |
+  <span>
+    <a [routerLink]=" ['./todo'] ">
+      Todo
+    </a>
+  </span>
+  |
+  <span>
+    <a [routerLink]=" ['./admin'] ">
+      Admin
+    </a>
+  </span>
+  |
+  <span>
+    <a [routerLink]=" ['./login'] ">
+      Login
+    </a>
+  </span>
+  |
+  <span>
+    <a [routerLink]=" ['./register'] ">
+      Register
+    </a>
+  </span>
+  |
+  <span>
+    <a [routerLink]=" ['./about'] ">
+      About
+    </a>
+  </span>
+</nav>
+
+<main>
+  <router-outlet></router-outlet>
+</main>
+
+<vu-chat></vu-chat>
+
+<pre class="app-state">this.appState.state = {{ appState.state | json }}</pre>
+
+<footer>
+  <span>{{ name }} by <a [href]="url">{{ author }}</a></span>
+  <div>
+    <a [href]="url">
+      <img [src]="appLogo" width="25%">
+    </a>
+  </div>
+</footer>
