+++ src/client/app/register/index.ts
@@ -1,2 +1,1 @@
-export * from './register.component';
-export * from './register-routing.module';
+export * from './register.module';
